+++ lib/model/om/BaseJournalComment.php
@@ -599,6 +599,11 @@ abstract class BaseJournalComment extends BaseObject  implements Persistent
 			$deleteQuery = JournalCommentQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(JournalCommentPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "journal_comments")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -642,6 +647,11 @@ abstract class BaseJournalComment extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(JournalCommentPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "journal_comments")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(JournalCommentPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -662,6 +672,11 @@ abstract class BaseJournalComment extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(JournalCommentPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "journal_comments")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(JournalCommentPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1387,6 +1402,26 @@ abstract class BaseJournalComment extends BaseObject  implements Persistent
 		return (string) $this->exportTo(JournalCommentPeer::DEFAULT_STRING_FORMAT);
 	}
 
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && JournalCommentPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return JournalCommentPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate("insert", $oUser);
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate("update", $oUser);
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate("delete", $oUser);
+	}
+
 	// extended_timestampable behavior
 	
 	/**
@@ -1405,7 +1440,7 @@ abstract class BaseJournalComment extends BaseObject  implements Persistent
 	 */
 	public function getCreatedAtTimestamp()
 	{
-		return $this->getCreatedAt('U');
+		return (int)$this->getCreatedAt('U');
 	}
 	
 	/**
@@ -1424,7 +1459,7 @@ abstract class BaseJournalComment extends BaseObject  implements Persistent
 	 */
 	public function getUpdatedAtTimestamp()
 	{
-		return $this->getUpdatedAt('U');
+		return (int)$this->getUpdatedAt('U');
 	}
 	
 	/**
